Update README.md
In our system, the employers (servee) can hire freelancers to do their work in different domains like software development, writing, data entry work and product design etc. As well as the employee (server) can also browse for job with a reward. So that this project provide benefits for both the users (servee / server) in an efficient way with a user friendly environment.